fix: address PR review comments for issue #250
