+++ backend/src/main/java/com/senior/spm/repository/SprintTrackingLogRepository.java
@@ -18,6 +18,8 @@ public interface SprintTrackingLogRepository extends JpaRepository<SprintTrackin
 
     List<SprintTrackingLog> findByGroupId(UUID groupId);
 
+    List<SprintTrackingLog> findByGroupIdAndAssigneeGithubUsernameAndPrMergedTrue(UUID groupId, String assigneeGithubUsername);
+
     List<SprintTrackingLog> findBySprintId(UUID sprintId);
 
     @Modifying
